Search term wrapper that allows for different capitalisation of term

Inspired by the varying capitalisation of "NCOV" during the coronavirus pandemic (e.g. ncov, nCoV, NCOV, nCOV), this function allows for all possible configurations of lower- and upper-case letters in your search term.

mx_caps(x)

Arguments

x
Search term to be formatted

Value

The input string is return, but with each non-space character repeated in lower- and upper-case, and enclosed in square brackets. For example, mx_caps("ncov") returns "[Nn][Cc][Oo][Vv]"

Examples

# \donttest{

query <- c("coronavirus", mx_caps("ncov"))

mx_search(mx_snapshot("6c4056d2cccd6031d92ee4269b1785c6ec4d555b"), query)

#> Using medRxiv snapshot - 2020-08-22 15:42

#> Warning: running command 'timedatectl' had status 1

#> Found 1856 record(s) matching your search.

#> # A tibble: 1,856 x 14
#>       ID title      abstract     authors     date       category  doi    version
#>    <int> <chr>      <chr>        <chr>       <date>     <chr>     <chr>    <int>
#>  1    88 Higher en… "Background… Ebbeling, … 2019-07-11 Nutrition 10.11…       1
#>  2   676 Resilienc… "Here we de… Lacaze, P.… 2019-10-22 Genetic … 10.11…       1
#>  3   501 Patterns … "We simulta… Lefort-Bes… 2019-09-27 Psychiat… 10.11…       1
#>  4  1371 Novel cor… "Since firs… Read, J. M… 2020-01-28 Infectio… 10.11…       2
#>  5  1426 Incubatio… "The geogra… Linton, N.… 2020-02-18 Infectio… 10.11…       2
#>  6  1430 Epidemiol… "Objective:… Jung, S.-M… 2020-02-03 Epidemio… 10.11…       2
#>  7  1465 Estimatin… "We estimat… Cao, Z.; Z… 2020-01-29 Infectio… 10.11…       1
#>  8  1446 The incub… "Currently,… Backer, J.… 2020-01-30 Infectio… 10.11…       2
#>  9  1475 Estimated… "Traveller … Gostic, K.… 2020-02-03 Epidemio… 10.11…       2
#> 10  1473 Risk for … "On January… Du, Z.; Wa… 2020-02-17 Epidemio… 10.11…       4
#> # … with 1,846 more rows, and 6 more variables: author_corresponding <chr>,
#> #   author_corresponding_institution <chr>, link_page <chr>, link_pdf <chr>,
#> #   license <chr>, published <chr>

# }